Change health-check to detect real listening port
This comes from https://github.com/moodlehq/moodle-docker/issues/200
where @terrycampbell detected that the recently implemented health
check for exttests was not working properly under moodle-docker.

And that's because, there, we change the exttests port to 9000
in order to workaround some limitations with podman (cannot have
2 containers publishing the same port, basically).

While more investigations happen with podman, we are changing the
check here to inspect the configuration and use the port that is
effectively being used.